Re-consolidate plan.js into execute.js so that Pull request diffs are cleaner
